requirements.txt
pandas==1.5.2
python-dotenv==0.21.0
SnakeMD==0.11.0
Telethon==1.26.0
